# auth_APP/docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Autenticación y Autorización — auth_APP
  description: |
    Servicio de autenticación, autorización y gestión de usuarios para el sistema de
    gestión de certificados SSL/TLS.

    **Autenticación híbrida**: soporta directorio corporativo LDAP/Active Directory con
    fallback automático a usuarios locales en base de datos.

    **Tokens**: los tokens se almacenan en cookies `httpOnly` (no accesibles por JavaScript).
    - `accessToken`: duración 15 minutos.
    - `refreshToken`: duración 7 días, usado para renovar el access token sin re-login.

    **Roles disponibles**: `admin`, `editor`, `auditor`, `viewer`.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo
  - url: https://tu-servidor.com
    description: Servidor de producción
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Renovar tokens
      description: |
        Renueva el `accessToken` usando el `refreshToken` almacenado en cookie.
        Si el `refreshToken` no está en cookie, también se acepta en el body como
        fallback.
      operationId: refresh
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Refresh token (alternativa a la cookie, solo si no hay cookie)
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Tokens renovados correctamente. Nuevas cookies establecidas.
          headers:
            Set-Cookie:
              description: Cookies `accessToken` y `refreshToken` renovadas
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Refresh token ausente o con formato inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Refresh token expirado o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/users/{id}/language:
    patch:
      tags:
        - users
      summary: Actualizar idioma preferido
      description: |
        Actualiza el idioma preferido del usuario y re-emite un nuevo par de tokens
        con el claim `lang` actualizado.

        **Autorización**: el propio usuario puede cambiar su idioma; los administradores
        pueden cambiarlo a cualquier usuario.

        Idiomas disponibles: `ca` (catalán, por defecto) y `es` (castellano).
      operationId: updateLanguage
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario cuyo idioma se quiere actualizar
          schema:
            type: string
          example: "42"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLanguageInput'
            example:
              language: es
      responses:
        '200':
          description: Idioma actualizado. Cookies renovadas con el nuevo claim `lang`.
          headers:
            Set-Cookie:
              description: |
                Cookies `accessToken` y `refreshToken` renovadas con `lang` actualizado
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateLanguageResponse'
        '400':
          description: Campo `language` ausente o valor no permitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: "Language must be one of: ca, es"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El usuario intenta cambiar el idioma de otro usuario sin ser admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FORBIDDEN
                  message: Access denied
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # Admin — Role management
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: |
        Cookie `accessToken` establecida por `/auth/login` o `/auth/refresh`.
        Se trata de una cookie `httpOnly`, `SameSite=Strict`, inaccesible por JavaScript.
  responses:
    Unauthorized:
      description: Token ausente, expirado o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Token de acceso requerido
...
  schemas:
...
    RefreshResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/UserPayload'
        message:
          type: string
          example: Tokens refreshed successfully
...
    UserPayload:
      type: object
      description: Payload incluido en los tokens JWT
      properties:
        userId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        username:
          type: string
          example: jperez
        lang:
          type: string
          description: Idioma preferido del usuario (`ca` o `es`)
          enum: [ca, es]
          example: ca
        email:
          type: string
          format: email
          example: jperez@empresa.com
        roles:
          type: array
          description: Roles del usuario (para la aplicación solicitada, o de todas si no se especificó)
          items:
            type: string
            enum: [admin, editor, auditor, viewer]
          example: [editor]
        applicationName:
          type: string
          description: Aplicación para la que se emitió el token
          example: secHTTPS_APP
...
    UpdateLanguageInput:
      type: object
      required:
        - language
      properties:
        language:
          type: string
          description: Nuevo idioma preferido
          enum: [ca, es]
          example: es
    UpdateLanguageResponse:
      type: object
      properties:
        message:
          type: string
          example: Idioma actualizado
        language:
          type: string
          enum: [ca, es]
          example: es
        accessToken:
          type: string
          description: Nuevo access token con `lang` actualizado (también en cookie)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Nuevo refresh token (también en cookie)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Código de error interno
              example: INVALID_CREDENTIALS
            message:
              type: string
              description: Descripción legible del error
              example: Credenciales incorrectas